+++ khorosjx/places/__init__.py
@@ -6,10 +6,10 @@
 :Example:        TBD
 :Created By:     Jeff Shurtliff
 :Last Modified:  Jeff Shurtliff
-:Modified Date:  07 Jan 2020
+:Modified Date:  14 Jan 2020
 """
 # Define all modules that will be imported with the "import *" method
 __all__ = ['base', 'blogs', 'spaces']
 
 # Always import the places base module
-from . import base
+from . import base, blogs, spaces